fix(model): clean up Responses request mapping
